heuristic rescale temp

+++ src/chip.h
@@ -7,15 +7,15 @@
 
 class Chip {
 public:
-    size_t x;
-    size_t y;
-    size_t width;
-    size_t height;
-    std::vector<std::pair<size_t, size_t>> bumps; // Stores bump coordinates as pairs of (x, y)
+    int x;
+    int y;
+    int width;
+    int height;
+    std::vector<std::pair<int, int>> bumps; // Stores bump coordinates as pairs of (x, y)
 
     // Constructor
     Chip() {};
-    Chip(size_t _x, size_t _y, size_t _width, size_t _height) {
+    Chip(int _x, int _y, int _width, int _height) {
         x = _x;
         y = _y;
         width = _width;
@@ -24,7 +24,7 @@ public:
         bumps.push_back({0, 0});
     }
 
-    void addBump(size_t b_x, size_t b_y) {
+    void addBump(int b_x, int b_y) {
         bumps.push_back({b_x, b_y});
     }
