+++ src/gcell.h
@@ -6,13 +6,13 @@
 
 class GCell {
 public:
-    size_t left_edge_capacity;
-    size_t bottom_edge_capacity;
+    int left_edge_capacity;
+    int bottom_edge_capacity;
     double cost;
 
     GCell()
         : left_edge_capacity(0), bottom_edge_capacity(0), cost(0) {}
-    GCell(size_t left_edge_capacity, size_t bottom_edge_capacity, double cost)
+    GCell(int left_edge_capacity, int bottom_edge_capacity, double cost)
         : left_edge_capacity(left_edge_capacity), bottom_edge_capacity(bottom_edge_capacity), cost(cost) {}
 };
