+++ src/input_reader.cpp
@@ -36,19 +36,19 @@ void read_grid_map(std::ifstream &file) {
     }
     routing_area_gheight = routing_area_height / grid_height;
     routing_area_gwidth  = routing_area_width  / grid_width;
-    edge_capacities[0] = std::vector<std::vector<size_t>>(routing_area_gheight, std::vector<size_t>(routing_area_gwidth, 0));
-    edge_capacities[1] = std::vector<std::vector<size_t>>(routing_area_gheight, std::vector<size_t>(routing_area_gwidth, 0));
-    edge_capacities[2] = std::vector<std::vector<size_t>>(routing_area_gheight, std::vector<size_t>(routing_area_gwidth, 0));
-    edge_capacities[3] = std::vector<std::vector<size_t>>(routing_area_gheight, std::vector<size_t>(routing_area_gwidth, 0));
+    edge_capacities[0] = std::vector<std::vector<int>>(routing_area_gheight, std::vector<int>(routing_area_gwidth, 0));
+    edge_capacities[1] = std::vector<std::vector<int>>(routing_area_gheight, std::vector<int>(routing_area_gwidth, 0));
+    edge_capacities[2] = std::vector<std::vector<int>>(routing_area_gheight, std::vector<int>(routing_area_gwidth, 0));
+    edge_capacities[3] = std::vector<std::vector<int>>(routing_area_gheight, std::vector<int>(routing_area_gwidth, 0));
 }
 
 void read_edge_capacity(std::ifstream &file) {
     std::string line;
-    size_t i = 0, j = 0;
+    int i = 0, j = 0;
     std::getline(file, line);
     while (std::getline(file, line)) {
         std::stringstream ss(line);
-        size_t left, bottom;
+        int left, bottom;
         ss >> left >> bottom;
         edge_capacities[BOTTOM][i][j] = bottom;
         edge_capacities[LEFT][i][j] = left;
@@ -74,6 +74,7 @@ void read_cost_file(std::ifstream &file) {
         std::stringstream ss(line);
         std::string type;
         ss >> type;
+        // std::cout << type << '\n';
         if (type == ".alpha") {
             ss >> alpha;
         } else if (type == ".beta") {
@@ -91,34 +92,37 @@ void read_cost_file(std::ifstream &file) {
         } else if (type == ".l") {
             std::vector<std::vector<double>> cur_layer_costs(routing_area_gheight, std::vector<double>(routing_area_gwidth));
             double cost;
-            for (size_t i = 0 ; i < routing_area_gheight ; i++) {
+            for (int i = 0 ; i < routing_area_gheight ; i++) {
                 std::getline(file, line);
                 std::stringstream ss(line);
-                for (size_t j = 0 ; j < routing_area_gwidth ; j++) {
+                for (int j = 0 ; j < routing_area_gwidth ; j++) {
                     ss >> cost;
                     cur_layer_costs[i][j] = cost;
                     gcell_cost_max = std::max(gcell_cost_max, cost);
                 }
             }
             layer_costs[current_layer++] = std::move(cur_layer_costs);
+            // printf("(%d, %d)\n", layer_costs[current_layer-1].size(), layer_costs[current_layer-1][0].size());
         }
     }
+
+    // printf("(%d, %d), (%d, %d)", layer_costs[0].size(), layer_costs[0][0].size(), layer_costs[1].size(), layer_costs[1][0].size());
     // layer1_gcells = std::vector<std::vector<GCell>>(routing_area_gheight, std::vector<GCell>(routing_area_gwidth));
     // layer2_gcells = std::vector<std::vector<GCell>>(routing_area_gheight, std::vector<GCell>(routing_area_gwidth));
     
-    // for (size_t i = 0 ; i < routing_area_gheight ; i++) {
-    //     for (size_t j = 0 ; j < routing_area_gwidth ; j++) {
-    //         size_t idx = i*routing_area_gwidth + j;
-    //         size_t left_cap  = edge_capacities[idx].first;
-    //         size_t right_cap = edge_capacities[idx].second;
+    // for (int i = 0 ; i < routing_area_gheight ; i++) {
+    //     for (int j = 0 ; j < routing_area_gwidth ; j++) {
+    //         int idx = i*routing_area_gwidth + j;
+    //         int left_cap  = edge_capacities[idx].first;
+    //         int right_cap = edge_capacities[idx].second;
     //         layer1_gcells[i][j] = GCell(left_cap, right_cap, layer1_costs[i][j]);
     //     }
     // }
-    // for (size_t i = 0 ; i < routing_area_gheight ; i++) {
-    //     for (size_t j = 0 ; j < routing_area_gwidth ; j++) {
-    //         size_t idx = i*routing_area_gwidth + j;
-    //         size_t left_cap  = edge_capacities[idx].first;
-    //         size_t right_cap = edge_capacities[idx].second;
+    // for (int i = 0 ; i < routing_area_gheight ; i++) {
+    //     for (int j = 0 ; j < routing_area_gwidth ; j++) {
+    //         int idx = i*routing_area_gwidth + j;
+    //         int left_cap  = edge_capacities[idx].first;
+    //         int right_cap = edge_capacities[idx].second;
     //         layer2_gcells[i][j] = GCell(left_cap, right_cap, layer2_costs[i][j]);
     //     }
     // }
